Fixed issue resolving username from user id
